requirements.txt
alpaca-trade-api>=3.0.0
python-telegram-bot>=20.7
pandas>=2.0.0
numpy==1.24.3
pandas-ta==0.3.14b0
python-dotenv>=1.0.0
yfinance>=0.2.36
pytest>=7.0.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
